# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mwango Brain API Documentation",
        "description": "Documentação das APIs RESTful do sistema Mwango Brain - Gestão de Eventos, Leads e Contactos (Laravel 13 & Neon PostgreSQL)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local / Dev"
        }
    ],
    "paths": {
...
        "/api/leads/{leadId}/interactions": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Registar nova interacção para uma lead específica",
                "operationId": "storeInteractionForLead",
                "parameters": [
                    {
                        "name": "leadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Interacção registada com sucesso"
                    }
                }
            }
        },
...
        "/api/leads/{id}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Obter detalhes da lead por ID",
                "operationId": "getLeadById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead encontrada"
                    },
                    "404": {
                        "description": "Lead não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Leads"
                ],
                "summary": "Atualizar lead",
                "operationId": "updateLead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead atualizada"
                    },
                    "404": {
                        "description": "Lead não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Remover lead",
                "operationId": "deleteLead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead removida"
                    }
                }
            }
        },
        "/api/leads/{id}/status": {
            "patch": {
                "tags": [
                    "Leads"
                ],
                "summary": "Atualizar estado da lead",
                "operationId": "updateLeadStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado da lead atualizado"
                    }
                }
            }
        },
...
        "/api/solutions/{id}": {
            "delete": {
                "tags": [
                    "Solutions"
                ],
                "summary": "Remover solução",
                "operationId": "deleteSolution",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solução removida"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Solutions"
                ],
                "summary": "Atualizar solução",
                "operationId": "updateSolution",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solução atualizada"
                    }
                }
            }
        },
...
        "/api/users/{id}/approve": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Aprovar conta de utilizador pendente",
                "operationId": "approveUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilizador aprovado"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Rejeitar ou remover utilizador",
                "operationId": "rejectUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilizador removido"
                    }
                }
            }
        },
...
    },
...
}